Update: Added maxSegmentsPerCell to initialisation of region
Since we are planning to use a three dimensional matrix to map
from every input to every segment it is necessary to use a
parameter that sets the maximal number of segements per cell.
Then it is possible to map every segment to the corresponding
cell and vice versa.
Not used segmental weights are set to 0