Update: Implemented moving averages
Added necessary variables for moving averages and added calculation
in the basal learning method. Implemented additionally a mask to
determine correctly predicted values and incorrectly predicted cells.
Values of incorrectly predicted are substracted, the others added.

Question: Is this the right way to go? Should another mechanism
be implemented here?