Update: Added mask that was not committed with previous commit
Now additionally a mask is implemented that was not previously
committed. It determines the correctly predicted cells. Values
for correctly predicted cells are added while incorrectly predicted
cells are subtracted.

Question. Is this the right way to? Should another punishment mechanism
be implemented? Could lead to an increment of incorrectly predicted
cells with the moving averages