Update: Treat apical and basal input equally in computation
To determine the cell's value the max function is applied
to the segment values. However, apical input is treated now
as additional information. Hence, they are added and then the max
value is taken. To make this a sound mathematical operation
it is required to include the bias value for the activation
only once. Thus, a flag can be passed to determine whether
the bias value is taken into account or not