Update: Reduced num of objects and added output
For efficiency reasons the number of objects are reduced.
So far, when having only a single object saved it can recognize it
immidiately whereas using more objects make it fail.
It seems like that the output does not change after passing different
sensations. Problem with not adapted L2 or bug?