Update: Implemented segment count for each cell
Instead of using the most activated cell we are now using the cell with the
least segments for learning new contexts. This is more efficient and
distributes the segments more evenly. Moreover, we are using now a
segment count meaning that unused segments that have been already
appended previously by a new row can be used.